cleanup and configs

+++ src/main/java/net/thep2wking/oedldoedlconstruct/content/modifier/ModifierWorldBreaker.java
@@ -6,37 +6,21 @@ import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraft.world.World;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.event.world.BlockEvent.BreakEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import net.thep2wking.oedldoedlconstruct.OedldoedlConstruct;
-import slimeknights.tconstruct.library.modifiers.ModifierAspect;
-import slimeknights.tconstruct.library.modifiers.ModifierTrait;
+import net.thep2wking.oedldoedlconstruct.api.ModSingleModifierBase;
+import net.thep2wking.oedldoedlconstruct.config.ConstructConfig;
 import slimeknights.tconstruct.library.utils.TagUtil;
 import slimeknights.tconstruct.library.utils.TinkerUtil;
 import slimeknights.tconstruct.library.utils.ToolHelper;
 
-public class ModifierWorldBreaker extends ModifierTrait {
-	public ModifierWorldBreaker() {
-		super("world_breaker", 0xc29ee5);
-		this.addAspects(new ModifierAspect.SingleAspect(this));
-		this.addItem("remoteBreaker");
+public class ModifierWorldBreaker extends ModSingleModifierBase {
+	public ModifierWorldBreaker(String identifier, int color, String modifierItemOreDict) {
+		super(identifier, color, modifierItemOreDict);
 		MinecraftForge.EVENT_BUS.register(this);
-	}
-
-	@Override
-	public String getLocalizedName() {
-		return new TextComponentTranslation("modifier." + OedldoedlConstruct.MODID + "." + this.identifier + ".name")
-				.getFormattedText();
-	}
-
-	@Override
-	public String getLocalizedDesc() {
-		return new TextComponentTranslation("modifier." + OedldoedlConstruct.MODID + "." + this.identifier + ".desc")
-				.getFormattedText();
 	}
 
 	@Override
@@ -45,7 +29,8 @@ public class ModifierWorldBreaker extends ModifierTrait {
 		BlockPos pos = event.getPos();
 		IBlockState state = world.getBlockState(pos);
 		Block block = state.getBlock();
-		if (state.getBlockHardness(world, pos) <= -1) {
+		if (state.getBlockHardness(world, pos) <= -1
+				&& ConstructConfig.CONTENT.MODIFIER.WORLD_BREAKER_MODIFIER_BREAKS_UNBREAKABLE) {
 			if (!world.isRemote) {
 				world.destroyBlock(pos, false);
 				world.setBlockToAir(pos);
@@ -67,7 +52,8 @@ public class ModifierWorldBreaker extends ModifierTrait {
 		ItemStack tool = player.getHeldItemMainhand();
 		if ((TinkerUtil.hasModifier(TagUtil.getTagSafe(tool), this.getIdentifier())
 				|| TinkerUtil.hasTrait(TagUtil.getTagSafe(tool), this.getIdentifier()))
-				&& ToolHelper.getCurrentDurability(tool) >= 1) {
+				&& ToolHelper.getCurrentDurability(tool) >= 1
+				&& ConstructConfig.CONTENT.MODIFIER.WORLD_BREAKER_MODIFIER_BREAKS_UNBREAKABLE) {
 			if (state.getBlockHardness(world, pos) <= -1 && !world.isRemote) {
 				if (!player.isCreative()) {
 					world.destroyBlock(pos, false);
